refactor: Adhere to coding standards
