+++ src/main/java/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package Tasks;
+package tasks;
 
 public class Deadline extends Task {
     private final String deadline;
