+++ src/main/java/tasks/Task.java
@@ -1,4 +1,4 @@
-package Tasks;
+package tasks;
 
 public abstract class Task {
     private final String name;
