+++ src/main/java/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package Tasks;
+package tasks;
 
 public class ToDo extends Task {
     public ToDo(String name) {
